refactor(utils): export readFile function from index
