+++ utils/index.ts
@@ -1,2 +1,4 @@
 export * from './blog_helpers'
 export * from './mailchimp_helper'
+export * from './readFile'
+export * from './getDynamicPath'
